refactor(#267 F8): colocate ssml unit tests with their submodules
The audit (#267 P1.F8): 660-LOC \`#[cfg(test)] mod tests\` block at the
bottom of \`tts/ssml/mod.rs\` mixed integration scenarios (full
\`<speak>...</speak>\` parse calls) with unit tests for sub-helpers
(\`parse_rate_value\`, \`Segment\` variant existence).

Move the unit tests next to the code they exercise — tests as
documentation work best when they're the first thing a reader hits
in the module:

- 8 \`parse_rate_value\` tests → \`tts/ssml/rate.rs::tests\`
- 2 \`Segment\` variant existence tests → \`tts/ssml/segment.rs::tests\`

Integration scenarios (parse-end-to-end) stay in \`mod.rs::tests\` for
this PR. Moving them to \`rust/tests/ssml_integration.rs\` is deferred —
it needs a careful pass over \`DEFAULT_BREAK\` visibility (currently
\`pub(super) const\`, accessible from submodule tests but not from a
top-level integration test).

Gates:
  cargo fmt --check                            clean
  cargo clippy --all-targets -- -D warnings    clean
  cargo test --lib tts::ssml                   45 / 45 passing

Refs #267
